migrate old tests and remove RefreshingManagedChannel and SafeShutdownManagedChannel
