--- gax-grpc/src/test/java/com/google/api/gax/grpc/SafeShutdownManagedChannelTest.java
@@ -1,176 +1,0 @@
-/*
- * Copyright 2019 Google LLC
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are
- * met:
- *
- *     * Redistributions of source code must retain the above copyright
- * notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above
- * copyright notice, this list of conditions and the following disclaimer
- * in the documentation and/or other materials provided with the
- * distribution.
- *     * Neither the name of Google LLC nor the names of its
- * contributors may be used to endorse or promote products derived from
- * this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package com.google.api.gax.grpc;
-
-import com.google.api.gax.grpc.testing.FakeMethodDescriptor;
-import io.grpc.CallOptions;
-import io.grpc.ClientCall;
-import io.grpc.ManagedChannel;
-import io.grpc.Metadata;
-import io.grpc.MethodDescriptor;
-import io.grpc.Status;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
-import org.mockito.Mockito;
-import org.mockito.stubbing.Answer;
-
-@RunWith(JUnit4.class)
-public class SafeShutdownManagedChannelTest {
-  // call should be allowed to complete and the channel should not shutdown
-  @Test
-  public void callShouldCompleteAfterCreation() {
-    final ManagedChannel underlyingChannel = Mockito.mock(ManagedChannel.class);
-
-    SafeShutdownManagedChannel safeShutdownManagedChannel =
-        new SafeShutdownManagedChannel(underlyingChannel);
-
-    // create a mock call when new call comes to the underlying channel
-    MockClientCall<String, Integer> mockClientCall = new MockClientCall<>(1, Status.OK);
-    MockClientCall<String, Integer> spyClientCall = Mockito.spy(mockClientCall);
-    Mockito.when(
-            underlyingChannel.newCall(
-                Mockito.<MethodDescriptor<String, Integer>>any(), Mockito.any(CallOptions.class)))
-        .thenReturn(spyClientCall);
-
-    Answer<?> verifyChannelNotShutdown =
-        invocation -> {
-          Mockito.verify(underlyingChannel, Mockito.never()).shutdown();
-          return invocation.callRealMethod();
-        };
-
-    // verify that underlying channel is not shutdown when clientCall is still sending message
-    Mockito.doAnswer(verifyChannelNotShutdown).when(spyClientCall).sendMessage(Mockito.anyString());
-
-    // create a new call on safeShutdownManagedChannel
-    @SuppressWarnings("unchecked")
-    ClientCall.Listener<Integer> listener = Mockito.mock(ClientCall.Listener.class);
-    ClientCall<String, Integer> call =
-        safeShutdownManagedChannel.newCall(
-            FakeMethodDescriptor.<String, Integer>create(), CallOptions.DEFAULT);
-
-    safeShutdownManagedChannel.shutdownSafely();
-    // shutdown is not called because there is still an outstanding call, even if it hasn't started
-    Mockito.verify(underlyingChannel, Mockito.after(200).never()).shutdown();
-
-    // start clientCall
-    call.start(listener, new Metadata());
-    // send message and end the call
-    call.sendMessage("message");
-    // shutdown is called because the outstanding call has completed
-    Mockito.verify(underlyingChannel, Mockito.atLeastOnce()).shutdown();
-  }
-
-  // call should be allowed to complete and the channel should not shutdown
-  @Test
-  public void callShouldCompleteAfterStarted() {
-    final ManagedChannel underlyingChannel = Mockito.mock(ManagedChannel.class);
-
-    SafeShutdownManagedChannel safeShutdownManagedChannel =
-        new SafeShutdownManagedChannel(underlyingChannel);
-
-    // create a mock call when new call comes to the underlying channel
-    MockClientCall<String, Integer> mockClientCall = new MockClientCall<>(1, Status.OK);
-    MockClientCall<String, Integer> spyClientCall = Mockito.spy(mockClientCall);
-    Mockito.when(
-            underlyingChannel.newCall(
-                Mockito.<MethodDescriptor<String, Integer>>any(), Mockito.any(CallOptions.class)))
-        .thenReturn(spyClientCall);
-
-    Answer<?> verifyChannelNotShutdown =
-        invocation -> {
-          Mockito.verify(underlyingChannel, Mockito.never()).shutdown();
-          return invocation.callRealMethod();
-        };
-
-    // verify that underlying channel is not shutdown when clientCall is still sending message
-    Mockito.doAnswer(verifyChannelNotShutdown).when(spyClientCall).sendMessage(Mockito.anyString());
-
-    // create a new call on safeShutdownManagedChannel
-    @SuppressWarnings("unchecked")
-    ClientCall.Listener<Integer> listener = Mockito.mock(ClientCall.Listener.class);
-    ClientCall<String, Integer> call =
-        safeShutdownManagedChannel.newCall(
-            FakeMethodDescriptor.<String, Integer>create(), CallOptions.DEFAULT);
-
-    // start clientCall
-    call.start(listener, new Metadata());
-    safeShutdownManagedChannel.shutdownSafely();
-    // shutdown is not called because there is still an outstanding call
-    Mockito.verify(underlyingChannel, Mockito.after(200).never()).shutdown();
-    // send message and end the call
-    call.sendMessage("message");
-    // shutdown is called because the outstanding call has completed
-    Mockito.verify(underlyingChannel, Mockito.atLeastOnce()).shutdown();
-  }
-
-  // Channel should shutdown after a refresh all the calls have completed
-  @Test
-  public void channelShouldShutdown() {
-    final ManagedChannel underlyingChannel = Mockito.mock(ManagedChannel.class);
-
-    SafeShutdownManagedChannel safeShutdownManagedChannel =
-        new SafeShutdownManagedChannel(underlyingChannel);
-
-    // create a mock call when new call comes to the underlying channel
-    MockClientCall<String, Integer> mockClientCall = new MockClientCall<>(1, Status.OK);
-    MockClientCall<String, Integer> spyClientCall = Mockito.spy(mockClientCall);
-    Mockito.when(
-            underlyingChannel.newCall(
-                Mockito.<MethodDescriptor<String, Integer>>any(), Mockito.any(CallOptions.class)))
-        .thenReturn(spyClientCall);
-
-    Answer<?> verifyChannelNotShutdown =
-        invocation -> {
-          Mockito.verify(underlyingChannel, Mockito.never()).shutdown();
-          return invocation.callRealMethod();
-        };
-
-    // verify that underlying channel is not shutdown when clientCall is still sending message
-    Mockito.doAnswer(verifyChannelNotShutdown).when(spyClientCall).sendMessage(Mockito.anyString());
-
-    // create a new call on safeShutdownManagedChannel
-    @SuppressWarnings("unchecked")
-    ClientCall.Listener<Integer> listener = Mockito.mock(ClientCall.Listener.class);
-    ClientCall<String, Integer> call =
-        safeShutdownManagedChannel.newCall(
-            FakeMethodDescriptor.<String, Integer>create(), CallOptions.DEFAULT);
-
-    // start clientCall
-    call.start(listener, new Metadata());
-    // send message and end the call
-    call.sendMessage("message");
-    // shutdown is not called because it has not been shutdown yet
-    Mockito.verify(underlyingChannel, Mockito.after(200).never()).shutdown();
-    safeShutdownManagedChannel.shutdownSafely();
-    // shutdown is called because the outstanding call has completed
-    Mockito.verify(underlyingChannel, Mockito.atLeastOnce()).shutdown();
-  }
-}
